+++ routes/arts.tsx
@@ -6,6 +6,7 @@ import "@utils/i18n/config.ts";
 import { PageProps } from "fresh";
 
 import ArtsSearch from "@islands/livesearch/ArtsSearch.tsx";
+import { DiscoverPaper } from "@islands/paper/DiscoverPaper.tsx";
 import Footer from "@islands/footer/Footer.tsx";
 import WaterDrop from "@islands/footer/WaterDrop.tsx";
 import { WorldMapPaper } from "@islands/paper/WorldMapPaper.tsx";
@@ -48,7 +49,14 @@ export default function ArtsPage(
         data-name="arts"
         class="relative flex-grow xl:max-h-screen scrollable xl:overflow-y-scroll custom-scrollbar overflow-hidden"
       >
+        {/* Post-it : découverte de l'Art */}
+        <div class="hidden 2xl:block absolute top-0 -left-16">
+          <DiscoverPaper />
+        </div>
+
+        {/* Recherche : œuvres d'art */}
         <ArtsSearch />
+
         {/* Post-it : lien vers la carte du Monde */}
         <div class="absolute top-1 right-0">
           <WorldMapPaper />